requirements.txt
asgiref==3.7.2
click==8.1.7
colorama==0.4.6
Django==5.0
django-redis==5.4.0
django-redis-cache==3.0.1
django-rq==2.10.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
redis==3.5.3
rq==1.15.1
sqlparse==0.4.4
tzdata==2023.4
